Update weigher and implement IMeasurableMemory in StreamingState
